% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plan.R
\name{add_exercise_stats}
\alias{add_exercise_stats}
\title{Add exercise stats}
\usage{
add_exercise_stats(data, exercise_stats)
}
\arguments{
\item{data}{PLACEHOLDER}

\item{exercise_stats}{PLACEHOLDER}
}
\description{
Internal helper function. Do not use!
}
