% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plan.R
\name{add_trainings}
\alias{add_trainings}
\title{Add Trainings}
\usage{
add_trainings(data, workouts)
}
\arguments{
\item{data}{PLACEHOLDER}

\item{workouts}{PLACEHOLDER}
}
\description{
Internal helper function. Do not use!
}
