% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plan.R
\name{create_daterange_tibble}
\alias{create_daterange_tibble}
\title{Create daterange tibble}
\usage{
create_daterange_tibble(start_date, time_span)
}
\arguments{
\item{start_date}{PLACEHOLDER}

\item{time_span}{PLACEHOLDER}
}
\description{
Internal helper function. Do not use!
}
