% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plan.R
\name{create_workout}
\alias{create_workout}
\title{Create a workout}
\usage{
create_workout(exercise, nset = 3, nrep = 5, step = 1)
}
\arguments{
\item{exercise}{Character vector of the exercises chosen.}

\item{nset}{Vector of number of sets per exercise. Must be either length \code{1} or equal to the length of \code{exercise}}

\item{nrep}{Vector of number of repetitions per set. Must be either length \code{1} or length of \code{exercise}.}

\item{step}{The step size for increasing the weight from last workout.}
}
\value{
A dataframe.
}
\description{
Choose exercises, sets and reps to create specific workouts.
}
\examples{
create_workout(exercise = c("benchpress", "squat", "overhead_press"), nset = 3, nrep = c(3, 3, 5))
}
