% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plan.R
\name{filter_weekdays}
\alias{filter_weekdays}
\title{Filter weekdays}
\usage{
filter_weekdays(data, weekdays)
}
\arguments{
\item{data}{PLACEHOLDER}

\item{weekdays}{PLACEHOLDER}
}
\description{
Internal helper function. Filter dates based on training days selected.
}
