% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strongr-package.R
\docType{package}
\name{strongr-package}
\alias{strongr}
\alias{strongr-package}
\title{strongr: Create Workout-Plans And Log Progress}
\description{
Create workout-plans form a set of input parameters. Log your progress with built in functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kdschneider.github.io/strongr}
}

}
\author{
\strong{Maintainer}: Konstantin Schneider \email{ks@konstantin-schneider.eu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
