% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plan.R
\name{create_plan}
\alias{create_plan}
\title{Create workout-plan}
\usage{
create_plan()
}
\value{
A dataframe containing all training sessions.
}
\description{
Use workouts/input parameters to create Workout-Plans.
}
